Updated to 1.0.0.0
An innumerable number of changes take place with the release 1.0.0.0. I have not updated Stratos' repo for a long time, and now I intend to make this a regular thing. Changes in 1.0.0.0 include the splitting of C.R.D..py (the bot file) into the modules under the "ext" folder, new tools for Event Horizon staff such as ;rslookup, updates to ;userinfo and much more.