Added filters
Configurable text filters have been added, that currently just delete any message containing a certain key word. There is a lot of work to be done here including on the user filter, all of which will be highly modular. Also changed the log function to no longer report when an embed updates.