Add ship builds and timeout
Added TimeOut (still experimental at this point), and ship builds to be displayed alongside ships in addition to database support. EOP is currently the only mod with some support and the database for that has been not included for obvious reasons. Various other changes included that I have not mentioned such as the EH temporary command for the 3000 member giveaway. Some functions also updated for additional safety against pinging everyone.